signal fix

+++ source/main.c
@@ -6,12 +6,13 @@
 /*   By: sven <sven@student.42.fr>                    +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/06/14 15:09:03 by psadeghi      #+#    #+#                 */
-/*   Updated: 2023/09/01 18:23:54 by svan-has      ########   odam.nl         */
+/*   Updated: 2023/09/01 20:33:44 by svan-has      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
+int	g_heredoc = 0;
 // void	blah(void)
 // {
 // 	system("leaks minishell");
@@ -28,7 +29,6 @@ char	*ft_readline(char *prompt, char **envp)
 	int		exit_status;
 	
 	// atexit(blah);
-	signals_init();
 	env = copy_environment_list(envp);
 	p_list = NULL;
 	syntax_check = 0;
@@ -36,6 +36,8 @@ char	*ft_readline(char *prompt, char **envp)
 	lst = NULL;
 	while (1)
 	{
+		g_heredoc = 0;
+		signals_parent();
 		lst = NULL;
 		unlink("here_doc");
 		line = readline(prompt);
@@ -43,16 +45,14 @@ char	*ft_readline(char *prompt, char **envp)
 		//if (!line || line[0] == '\0')
 		if (!new)
 		{
-			//printf("\b\b \n");
 			free(line);
 			exit(1);
 		}
-		if (!new || new[0] == '\0')
+		else if (new[0] == '\0')
 		{
 			free(line);
 			free(new);
-			line = readline(prompt);
-			new = ft_strtrim(line, " ");
+			continue ;
 		}
 		else
 		{
@@ -64,6 +64,14 @@ char	*ft_readline(char *prompt, char **envp)
 				// 	printf("this is the head\n");
 				expansion(&lst, &env, exit_status);
 				lst = make_parser(&lst, &p_list);
+				if (g_heredoc == 1)
+				{
+					// free(new);
+					// free(line);
+					// free_tokens(lst);
+					// free_llist(&p_list);
+					continue ;
+				}
 				execution(&p_list, &env, &exit_status);
 				unlink("here_doc");
 				// ft_putstr_fd("Return code: ", 1);
